+++ Jagerts.Kaleid.Application/TestScene.cpp
@@ -122,7 +122,7 @@ void TestScene::Render()
 	this->_app->GetWindow(0)->GetSize(&width, &height);
 	this->_camera.SetAspectRatio((float)width / height);
 	this->_renderer->Clear(0.0f, 0.0f, 0.0f, 1.0f);
-	this->_renderer->SetViewport(0.0f, 0.0f, width, height);
+	this->_renderer->SetViewport(0.0f, 0.0f, (float)width, (float)height);
 
 	// Render Scene
 	Matrix4F mvp = this->_camera.GetProjectionMatrix() * this->_camera.GetViewMatrix() * this->_cube.GetTransform()->GetModelMatrix();
